activatVC-platform/activatVC-agents-platform/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.0

# Database
sqlalchemy==2.0.35
alembic==1.13.3
asyncpg==0.29.0

# Pydantic
pydantic==2.9.2
pydantic-settings==2.5.2

# Async HTTP
httpx==0.27.2

# LLM providers
openai==1.50.0
anthropic==0.34.0
google-generativeai==0.8.0

# Document parsing
python-docx==1.1.2
markdown==3.7
PyMuPDF==1.24.10

# Export
reportlab>=4.0.0
xhtml2pdf>=0.2.11

# Utilities
python-multipart==0.0.12
tenacity==9.0.0
python-dotenv==1.0.1

# Tests
pytest==8.3.3
pytest-asyncio==0.24.0
